LayoutTests/platform/ios/imported/w3c/web-platform-tests/intersection-observer/iframe-no-root-with-wrapping-scroller-expected.txt



PASS Observer with the implicit root; target in a same-origin iframe. 
PASS First rAF. 
PASS document.scrollingElement.scrollTop = 200 
FAIL iframe.contentDocument.scrollingElement.scrollTop = 250 assert_equals: entries.length expected 2 but got 1
FAIL document.scrollingElement.scrollTop = 100 assert_equals: entries.length expected 3 but got 1
